Add mirror commit fetching
